requirements.txt
GitPython==2.1.3
Flask==0.12.1
